+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](https://semver.org/)
 
+
+## [Unreleased]
+
+### Changed
+
+- Change plugin labels from UUID[n] to UUIDv][n] 
+
 ## [3.0.0b1] 2026-06-19
 
 ### Changed